Move 3 misfiled EF configuration classes to owning section folders (#1178)
* chore: relocate misfiled EF configurations to owning section folders

Moves three EF configuration classes into their owning section's
Configurations/<Section>/ folder and updates namespaces to match:

- AccountMergeRequestConfiguration: Profiles -> Users
- EventParticipationConfiguration: Shifts -> Users
- GoogleResourceConfiguration: Configurations root -> GoogleIntegration

Pure relocation, no schema or behavior change. Moving GoogleResourceConfiguration
into a named section folder made the HUM0024 cross-section-EF-join analyzer newly
able to classify it, which surfaced a pre-existing (previously invisible)
GoogleIntegration -> Teams navigation join. Marked [Grandfathered("HUM0024", ...)]
consistent with the other pre-existing HUM0024 violators, per
docs/architecture/peters-hard-rules.md tech debt handling.

Fixes nobodies-collective/Humans#981


* chore: update namespaces for relocated EF configurations

Follow-up to a344f5c97, which moved the three files but left the
namespace/using edits uncommitted. Adds the actual:

- AccountMergeRequestConfiguration: namespace .Profiles -> .Users
- EventParticipationConfiguration: namespace .Shifts -> .Users
- GoogleResourceConfiguration: namespace Configurations -> Configurations.GoogleIntegration
  + [Grandfathered("HUM0024", ...)] (see PR description)

Fixes nobodies-collective/Humans#981


---------
